+++ src/components/DatePicker/datePickerChangelogData.ts
@@ -10,27 +10,29 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-03-14",
+    version: "3.0.0",
     type: "Update",
     affects: ["Styles"],
     notes: ["Chakra 2.8 update."],
   },
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-02-22",
+    version: "2.1.6",
     type: "Update",
     affects: ["Styles"],
     notes: [
-      "Fixes bug where month and year calendars were rendering vertically rather than horizontally.",
+      "Fixes the range calendar style for the month and year calendars.",
+      "Fixes minor typo in the Initial Date Values section.",
     ],
   },
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-01-04",
+    version: "2.1.4",
     type: "Update",
-    affects: ["Functionality"],
+    affects: ["Functionality", "Styles"],
     notes: [
+      "Fixes bug where month and year calendars were rendering vertically rather than horizontally.",
       "Updates the `initialDate` and `initialDateTo` props to accept an empty string and adds optional `placeholder` and `placeholderTo` props.",
     ],
   },