+++ src/components/Icons/iconChangelogData.ts
@@ -10,11 +10,18 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-03-14",
+    version: "3.0.0",
     type: "Update",
     affects: ["Styles"],
     notes: ["Chakra 2.8 update."],
+  },
+  {
+    date: "2024-02-05",
+    version: "2.1.5",
+    type: "Update",
+    affects: ["Documentation", "Functionality"],
+    notes: ["Added the navigationMoreVert (vertical ellipsis) icon."],
   },
   {
     date: "Prerelease",