+++ src/components/NewsletterSignup/newsletterSignupChangelogData.ts
@@ -10,11 +10,14 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-03-14",
+    version: "3.0.0",
     type: "Update",
-    affects: ["Styles"],
-    notes: ["Chakra 2.8 update."],
+    affects: ["Styles", "Functionality"],
+    notes: [
+      "Chakra 2.8 update.",
+      "Updated the email field error message to following NYPL recommendations and use more direct language.",
+    ],
   },
   {
     date: "2023-12-07",