+++ src/theme/components/tabs.ts
@@ -17,6 +17,7 @@ const { defineMultiStyleConfig, definePartsStyle } =
 
 const tablist = {
   borderColor: "ui.black",
+  width: "max-content",
 };
 const tab = {
   background: "transparent",
@@ -130,13 +131,18 @@ const tabpanels = {
 const carouselParent = {
   position: { base: "absolute", md: "relative" },
   left: { base: "50px", md: "auto" },
-  paddingBottom: { base: "5px", md: "0" },
-  paddingStart: { base: "4px", md: "0" },
-  paddingEnd: "0",
-  paddingTop: { base: "4px", md: "0" },
-  right: { base: "52px", md: "auto" },
-  top: { base: "4px", md: "0" },
-  overflowX: { base: "hidden", md: "visible" },
+  paddingBottom: { base: "5px", md: "4px" },
+  paddingStart: "4px",
+  paddingEnd: "4px",
+  paddingTop: { base: "4px" },
+  right: { base: "52px", md: "4px" },
+  top: { base: "4px" },
+  scrollbarWidth: "none",
+  "::-webkit-scrollbar": {
+    display: "none",
+  },
+  overflowY: "hidden",
+  overflowX: "scroll",
 };
 
 const CustomTabs = defineMultiStyleConfig({