+++ src/theme/index.ts
@@ -13,11 +13,12 @@ import AudioPlayer from "./components/audioPlayer";
 import Breadcrumb from "./components/breadcrumb";
 import Button from "./components/button";
 import ButtonGroup from "./components/buttonGroup";
-import Card from "./components/card";
 import Checkbox from "./components/checkbox";
 import ComponentWrapper from "./components/componentWrapper";
 import CheckboxGroup from "./components/checkboxGroup";
 import CustomSelect from "./components/select";
+import CustomSlider from "./components/slider";
+import CustomTable from "./components/customTable";
 import DatePicker from "./components/datePicker";
 import FeedbackBox from "./components/feedbackBox";
 import FeaturedContent from "./components/featuredContent";
@@ -32,9 +33,10 @@ import Label from "./components/label";
 import Link from "./components/link";
 import List from "./components/list";
 import Logo from "./components/logo";
+import Menu from "./components/menu";
 import Modal from "./components/modal";
 import MultiSelect from "./components/multiSelect";
-import MultiSelectMenuButton from "./components/multiSelectMenuButton";
+import MultiSelectItemsCountButton from "./components/multiSelectItemsCountButton";
 import NewsletterSignup from "./components/newsletterSignup";
 import Notification from "./components/notification";
 import NotificationContent from "./components/notificationContent";
@@ -43,17 +45,16 @@ import Pagination from "./components/pagination";
 import ProgressIndicator from "./components/progressIndicator";
 import Radio, { RadioWrapper } from "./components/radio";
 import RadioGroup from "./components/radioGroup";
+import ReservoirCard from "./components/card";
 import { ReservoirImage, ReservoirImageWrapper } from "./components/image";
 import SearchBar from "./components/searchBar";
-import { Skeleton, SkeletonLoader } from "./components/skeletonLoader";
+import SkeletonLoader from "./components/skeletonLoader";
 import SkipNavigation from "./components/skipNavigation";
-import CustomSlider from "./components/slider";
 import StatusBadge from "./components/statusBadge";
 import StructuredContent from "./components/structuredContent";
 import StyledList from "./components/styledList";
 import SocialMediaLinks from "./components/socialmedialinks";
 import Tabs from "./components/tabs";
-import CustomTable from "./components/customTable";
 import TagSetStyles from "./components/tagSet";
 import TemplateStyles from "./components/template";
 import Text from "./components/text";
@@ -99,7 +100,6 @@ const theme: any = {
     CustomBreadcrumb: Breadcrumb,
     CustomButton: Button,
     ButtonGroup,
-    ...Card,
     Checkbox,
     CheckboxGroup,
     ComponentWrapper,
@@ -117,9 +117,10 @@ const theme: any = {
     Link,
     ReservoirList: List,
     Logo,
+    Menu,
     Modal,
     MultiSelect,
-    MultiSelectMenuButton,
+    MultiSelectItemsCountButton,
     Notification,
     NotificationContent,
     NotificationHeading,
@@ -129,11 +130,11 @@ const theme: any = {
     Radio,
     RadioGroup,
     RadioWrapper,
+    ...ReservoirCard,
     ReservoirIcon: Icon,
     ReservoirImage,
     ReservoirImageWrapper,
     SearchBar,
-    Skeleton,
     SkeletonLoader,
     SkipNavigation,
     CustomSlider,